Adjusting responsiveness
