+++ styles/pages/index.js
@@ -6,52 +6,6 @@ const useStyles = makeStyles(theme => ({
     flexDirection: 'column'
     // height: '100%'
   },
-  banner: {
-    position: 'relative',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    background: 'url(/header.jpg)',
-    backgroundRepeat: 'no-repeat',
-    backgroundSize: 'cover',
-    height: '30%'
-    // marginBottom: theme.spacing(4)
-  },
-  container: {
-    background: 'transparent',
-    position: 'relative',
-    textAlign: 'center',
-    color: '#FFF',
-    zIndex: 2,
-    marginTop: '0'
-  },
-  title: {
-    fontFamily: 'Oxanium',
-    fontWeight: 100,
-    fontSize: 50,
-    margin: 0,
-    [theme.breakpoints.down('sm')]: {
-      fontSize: 48,
-      margin: `${theme.spacing(5)}px ${theme.spacing(2)}px`
-    },
-
-    textShadow: 'black 0.1em 0.1em 0.2em'
-  },
-  titleWrapper: {
-    minHeight: 240,
-    [theme.breakpoints.up('sm')]: {
-      display: 'flex',
-      justifyContent: 'center',
-      alignItems: 'center'
-    }
-  },
-  socialWrapper: {
-    color: '#fff',
-    position: 'absolute',
-    right: 0,
-    bottom: 0,
-    zIndex: 2
-  },
   interfaceContainer: {
     gap: 16,
 
@@ -63,9 +17,10 @@ const useStyles = makeStyles(theme => ({
   main: {
     position: 'relative',
     zIndex: 2,
-    height: '100%',
     display: 'flex',
-    alignItems: 'center'
+    alignItems: 'center',
+    height: '100%',
+    padding: '20px 0'
   },
   titleItem: {
     fontFamily: 'Oxanium',
